remote_agent_agp/requirements.txt
langgraph==0.3.2
langchain==0.3.19
langchain_openai==0.3.8
langgraph-sdk==0.1.53
fastapi[standard]==0.115.11
requests==2.32.3
starlette==0.46.0
uvicorn==0.34.0
packaging==24.2
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
uvicorn==0.34.0
requests==2.32.3
langgraph-cli[inmem]==0.1.74
python-json-logger==3.3.0
agp-bindings==0.1.14
agp-api==0.0.6
langsmith==0.3.18
# If you need to debug langgraph studio
# debugpy==1.8.12
